+++ examples/ODE/bench_sodu_omp.cpp
@@ -1,9 +1,9 @@
 #include <iostream>
 #include "sodu.hpp"
 
-float koef_matrix_settings ( unsigned j, unsigned i, unsigned n )
+double koef_matrix_settings ( unsigned j, unsigned i, unsigned n )
     {
-    float value;
+    double value;
     if ( j==i ) value = -4.f;
     else if ( j>0 && ( ( j-1 ) == i ) ) value = 1.f;
     else if ( j< ( n-1 ) && ( ( j+1 ) == i ) ) value = 1.f;
@@ -11,24 +11,24 @@ float koef_matrix_settings ( unsigned j, unsigned i, unsigned n )
     return value;
     }
 
-simfor::matrix<float> odu_matrix_create ( unsigned n )
+simfor::matrix<double> odu_matrix_create ( unsigned n )
     {
-    simfor::matrix<float> M ( n, n );
+    simfor::matrix<double> M ( n, n );
     for ( int i = 0; i < n; i++ )
         for ( int j = 0; j < n; j++ )
             M ( i, j ) = koef_matrix_settings ( j,i,n );
     return M;
     }
 
-float integral_function ( float x, int number )
+double integral_function ( double x, int number )
     {
     return exp ( -1 * powf ( x - number, 2 ) );
     }
 
-float integral_result ( float x, int number )
+double integral_result ( double x, int number )
     {
-    float integral = 0;
-    float gauss_x[7] = {-1.f,
+    double integral = 0;
+    double gauss_x[7] = {-1.f,
                         - sqrtf ( 5. / 11. + 2. / 11. * sqrtf ( 5. / 3. ) ),
                         - sqrtf ( 5. / 11. - 2. / 11. * sqrtf ( 5. / 3. ) ), 0.0,
                         sqrtf ( 5. / 11. - 2. / 11. * sqrtf ( 5. / 3. ) ),
@@ -51,9 +51,9 @@ float integral_result ( float x, int number )
     return integral;
     }
 
-simfor::vector<float> initial_cond ( float a, int n )
+simfor::vector<double> initial_cond ( double a, int n )
     {
-    simfor::vector<float> y0 ( n );
+    simfor::vector<double> y0 ( n );
     for ( unsigned i = 0; i < n; ++i )
         y0 ( i ) = integral_result ( a, i );
     return y0;
@@ -62,10 +62,10 @@ simfor::vector<float> initial_cond ( float a, int n )
 int main ( int argc, char* argv [] )
     {
     int n = atoi ( argv [ 1 ] ), p = atoi ( argv [ 2 ] );
-    float a=0, b=1, h;
+    double a=0, b=1, h;
     h = ( b - a ) / n;
-    simfor::matrix<float> F = odu_matrix_create ( n ), Y;
-    simfor::vector<float> x0 = initial_cond ( 1, n );
+    simfor::matrix<double> F = odu_matrix_create ( n ), Y;
+    simfor::vector<double> x0 = initial_cond ( 1, n );
 
     omp_set_num_threads ( p );
